Dependent entities directly
