+++ lib/src/behaviors/card_creatable.dart
@@ -1,8 +1,6 @@
 import 'package:meta/meta.dart';
 import '../entities/authentication_session.dart';
 import '../entities/new_card_text.dart';
-export '../entities/authentication_session.dart';
-export '../entities/new_card_text.dart';
 
 abstract class CardCreatable {
   Future<void> add({
