+++ lib/src/behaviors/card_deletable.dart
@@ -1,8 +1,6 @@
 import 'package:meta/meta.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
-export '../entities/authentication_session.dart';
-export '../entities/card.dart';
 
 abstract class CardDeletable {
   Future<void> deleteCard({
