+++ lib/src/behaviors/card_subscribable.dart
@@ -2,8 +2,6 @@ import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
-export '../entities/authentication_session.dart';
-export '../entities/card.dart';
 
 abstract class CardSubscribable {
   ValueObservable<List<Card>> subscribeCards({
