+++ lib/src/behaviors/in_queue_card_subscribable.dart
@@ -2,8 +2,6 @@ import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
-export '../entities/authentication_session.dart';
-export '../entities/card.dart';
 
 abstract class InQueueCardSubscribable {
   Observable<List<Card>> subscribeInQueueCards(
