+++ lib/src/behaviors/logger_user_state_settable.dart
@@ -1,6 +1,5 @@
 import 'package:meta/meta.dart';
 import '../entities/user.dart';
-export '../entities/user.dart';
 
 abstract class LoggerUserStateSettable {
   Future<void> setUserState({@required User user});
