+++ lib/src/behaviors/review_subscribable.dart
@@ -3,9 +3,6 @@ import 'package:rxdart/rxdart.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
 import '../entities/review.dart';
-export '../entities/authentication_session.dart';
-export '../entities/card.dart';
-export '../entities/review.dart';
 
 abstract class ReviewSubscribable {
   Observable<List<Review>> subscribeReviewsInCard({
