+++ lib/src/behaviors/signable.dart
@@ -1,6 +1,5 @@
 import 'package:meta/meta.dart';
 import '../entities/authentication_session.dart';
-export '../entities/authentication_session.dart';
 
 abstract class Signable {
   Future<AuthenticationSession> restoreSession();
