+++ lib/src/behaviors/user_data_gettable.dart
@@ -1,6 +1,5 @@
 import 'package:meta/meta.dart';
 import '../entities/authentication_session.dart';
-export '../entities/authentication_session.dart';
 
 abstract class UserDataGettable {
   Future<UserData> getUserData({@required AuthenticationSession session});
