+++ lib/src/blocs/authentication_bloc.dart
@@ -1,11 +1,11 @@
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
-import '../entities/user.dart';
 import '../behaviors/logger_user_state_settable.dart';
 import '../behaviors/sign_in_out_loggable.dart';
 import '../behaviors/signable.dart';
 import '../behaviors/user_data_gettable.dart';
-export '../entities/user.dart';
+import '../entities/authentication_session.dart';
+import '../entities/user.dart';
 
 abstract class AuthenticationBloc {
   AuthenticationSession get session;
