+++ lib/src/blocs/card_delete_bloc.dart
@@ -1,7 +1,8 @@
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 import '../behaviors/card_deletable.dart';
-export '../behaviors/card_deletable.dart' show Card;
+import '../entities/authentication_session.dart';
+import '../entities/card.dart';
 
 abstract class CardDeleteBloc {
   Card get card;
