+++ lib/src/blocs/card_detail_bloc.dart
@@ -1,11 +1,9 @@
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
+import '../behaviors/review_subscribable.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
 import '../entities/review.dart';
-import '../behaviors/review_subscribable.dart';
-export '../entities/card.dart';
-export '../entities/review.dart';
 
 abstract class CardDetailBloc {
   Card get card;
