+++ lib/src/blocs/card_list_bloc.dart
@@ -1,9 +1,8 @@
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
+import '../behaviors/card_subscribable.dart';
 import '../entities/authentication_session.dart';
 import '../entities/card.dart';
-import '../behaviors/card_subscribable.dart';
-export '../entities/card.dart';
 
 abstract class CardListBloc {
   ValueObservable<List<Card>> get cards;
