+++ lib/src/blocs/card_review_bloc.dart
@@ -1,12 +1,11 @@
 import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
-import '../entities/card.dart';
-import '../entities/review_certainty.dart';
 import '../behaviors/card_reviewable.dart';
 import '../behaviors/card_review_loggable.dart';
 import '../behaviors/in_queue_card_subscribable.dart';
-export '../entities/card.dart';
-export '../entities/review_certainty.dart';
+import '../entities/authentication_session.dart';
+import '../entities/card.dart';
+import '../entities/review_certainty.dart';
 
 abstract class CardReviewBloc {
   ValueObservable<bool> isInitialized;
