+++ lib/src/implementations/firestore_user_repository.dart
@@ -2,6 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:meta/meta.dart';
 import '../behaviors/introduction_finishable.dart';
 import '../behaviors/user_data_gettable.dart';
+import '../entities/authentication_session.dart';
 
 class FirestoreUserRepository
     implements IntroductionFinishable, UserDataGettable {
