+++ lib/src/widgets/authentication_navigator/authentication_navigator.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import '../../blocs/authentication_bloc.dart';
+import '../../entities/user.dart';
 import './signed_in_route.dart';
 import './signed_out_route.dart';
 import './undecided_route.dart';
