+++ lib/src/widgets/card_delete_confirm_dialog.dart
@@ -4,6 +4,7 @@ import 'package:provider/provider.dart';
 import 'package:saraka/constants.dart';
 import '../blocs/authentication_bloc.dart';
 import '../blocs/card_delete_bloc.dart';
+import '../entities/card.dart';
 import './disappearable_builder.dart';
 import './fancy_popup_dialog.dart';
 import './processable_fancy_button.dart';
