+++ lib/src/widgets/dashboard_screen/card_list_button.dart
@@ -3,6 +3,7 @@ import 'package:flutter_icons/flutter_icons.dart';
 import 'package:provider/provider.dart';
 import 'package:saraka/constants.dart';
 import '../../blocs/card_list_bloc.dart';
+import '../../entities/card.dart';
 
 class CardListButton extends StatelessWidget {
   @override
