+++ lib/src/widgets/dashboard_screen/ready_card_length_text.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:saraka/constants.dart';
 import '../../blocs/card_list_bloc.dart';
+import '../../entities/card.dart';
 
 class ReadyCardLengthText extends StatelessWidget {
   @override
