+++ lib/src/widgets/review_screen/stacked_card.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart' hide Card;
 import 'package:flutter_icons/flutter_icons.dart';
 import 'package:provider/provider.dart';
 import 'package:saraka/constants.dart';
-import '../../blocs/card_review_bloc.dart';
 import '../../blocs/synthesizer_bloc.dart';
+import '../../entities/card.dart';
 
 @immutable
 class StackedCard extends StatelessWidget {
