+++ lib/src/widgets/signed_in_navigator/signed_in_navigator.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import '../../blocs/authentication_bloc.dart';
+import '../../entities/user.dart';
 import './card_list_route.dart';
 import './dashboard_route.dart';
 import './introduction_route.dart';
